+++ R/netmatch.r
@@ -1,5 +1,6 @@
 #' Matching Estimators with Network Data
 #'
+#' \strong{WARNING}: This function is still in development and has not been tested throughly.
 #' Following Aral et al. (2009), \code{netmatch} computes matching
 #' estimators for network data. The function \code{netmatch_prepare}, which
 #' prepares the data to be used with \code{\link[MatchIt:matchit]{matchit}} from