+++ R/struct_test.R
@@ -208,6 +208,15 @@ c.diffnet_struct_test <- function(..., recursive=FALSE) {
 #' @rdname struct_test
 print.diffnet_struct_test <- function(x, ...) {
 
+  # Neat column printing
+  netcol <- function(obs, expe, pval) {
+    txt <- paste0(sprintf("  %10.4f  %10.4f  %10.4f", obs, expe, pval), collapse="\n")
+    paste(
+      sprintf("  %10s  %10s  %10s", "observed", "expected", "p.val"),
+      txt, sep = "\n"
+    )
+  }
+
   with(x,  {
     nsim <- ifelse(!is.na(R), R, 0)
 
@@ -217,9 +226,7 @@ print.diffnet_struct_test <- function(x, ...) {
         "# of time periods : ", formatC(nslices(x$graph), digits = 0, format = "f", big.mark = ","),"\n",
         paste(rep("-",80), collapse=""),"\n",
         " H0: E[beta(Y,G)|G] - E[beta(Y,G)] = 0 (no structure dependency)\n",
-        "   E[beta(Y,G)|G] (observed) = ", paste0(sprintf("%12.4f",t0), collapse=", "), "\n",
-        "   E[beta(Y,G)] (expected)   = ", paste0(sprintf("%12.4f",mean_t), collapse=", "), "\n",
-        "   p-value                   = ", paste0(sprintf("%12.4f", p.value), collapse=", "),"\n",
+        netcol(t0, mean_t, p.value),
         sep="")
   })
   invisible(x)
@@ -228,6 +235,8 @@ print.diffnet_struct_test <- function(x, ...) {
 #' @export
 #' @param b0 Character scalar. When \code{annotated=TRUE}, label for the value of \code{b0}.
 #' @param b Character scalar. When \code{annotated=TRUE}, label for the value of \code{b}.
+#' @param ask Logical scalar. When \code{TRUE}, asks the user to type \code{<Enter>} to see each plot (as
+#' many as statistics where computed).
 #' @rdname struct_test
 hist.diffnet_struct_test <- function(
   x,
@@ -237,23 +246,44 @@ hist.diffnet_struct_test <- function(
   annotated=TRUE,
   b0=expression(atop(plain("") %up% plain("")), t[0]),
   b =expression(atop(plain("") %up% plain("")), t[]),
+  ask = TRUE,
   ...) {
 
-  out <- hist(x$boot$t,  breaks=breaks, plot=FALSE)
-  ran <- range(out$mids)
-  if (annotated) {
-    mt <- mean(x$boot$t, na.rm=TRUE)
-    ran <- range(c(ran, mt, x$boot$t0))
-    hist(x$boot$t, breaks=breaks, main=main, xlab=xlab, xlim = ran, ...)
-  } else {
-    hist(x$boot$t, breaks=breaks, main=main, xlab=xlab, ...)
+  # Par parameters
+  oldpar <- par(no.readonly = TRUE)
+  on.exit(par(oldpar))
+  par(ask=ask)
+
+  out <- vector("list", ncol(x$boot$t))
+  for (i in 1:ncol(x$boot$t)) {
+
+    # Computing the histogram and its range
+    out[[i]] <- hist(x$boot$t[,i],  breaks=breaks, plot=FALSE)
+    ran      <- range(out[[i]]$mids, na.rm = TRUE)
+
+    if (annotated) {
+
+      # Annotated version (pretty)
+      mt <- mean(x$boot$t[,i], na.rm=TRUE)
+      ran <- range(c(ran, mt, x$boot$t0[i]), na.rm=TRUE)
+      hist(x$boot$t[,i], breaks=breaks, main=main, xlab=xlab, xlim = ran, ...)
+
+    } else {
+
+      # Not annotated version
+      hist(x$boot$t[,i], breaks=breaks, main=main, xlab=xlab, ...)
+
+    }
+
+    # Adding margin note
+    if (annotated) {
+
+      mtext(b0, side = 1, at=x$boot$t0[i])
+      mtext(b, side = 1, at=mt)
+
+    }
   }
 
-  # Adding margin note
-  if (annotated) {
-    mtext(b0, side = 1, at=x$boot$t0)
-    mtext(b, side = 1, at=mt)
-  }
   invisible(out)
 }
 # #' @rdname struct_test