+++ tests/testthat/test-bootnet.r
@@ -16,6 +16,7 @@ context("Network bootstrap")
 #   unit= "relative"
 # )
 
+# ------------------------------------------------------------------------------
 test_that("Filling zeros", {
 
   set.seed(123)
@@ -23,16 +24,17 @@ test_that("Filling zeros", {
   g <- rgraph_ba(t = n-1, self=FALSE, m = 1)
 
   set.seed(1)
-  ans0 <- bootnet(g, function(w, i, ...) length(w@x), R=100)
+  ans0 <- bootnet(g, function(w, i, ...) length(w@x), R=100,
+                  resample.args = list(self=FALSE, useR=FALSE))
 
   set.seed(1)
   ans1 <- bootnet(g, function(w, i, ...) length(w@x), R=100,
-                  resample.args = list(self=FALSE, useR=FALSE))
+                  resample.args = list(self=FALSE, useR=TRUE))
 
   expect_equal(ans1[-length(ans0)], ans1[-length(ans1)])
 })
 
-
+# ------------------------------------------------------------------------------
 test_that("Methods", {
   # Generating the data
   set.seed(1291)
@@ -62,6 +64,19 @@ test_that("Methods", {
 
   expect_equivalent(ans0, ans1)
   expect_equivalent(ans0, ans2)
+})
+
+# ------------------------------------------------------------------------------
+test_that("diffnet_bootnet methods", {
+  set.seed(1222)
+  x <- rgraph_ba(t=19, m=1)
+  ans <- bootnet(x, function(g,...) mean(dgr(g)), R=50)
+
+  expect_output(print(ans), "Network Bootstrap")
+  expect_silent(hist(ans, ask=FALSE))
+  expect_s3_class(c(ans, ans), "diffnet_bootnet")
+  expect_output(print(c(ans, ans)), ": 100")
+
 })
 # rn(list=ls())
 # library(microbenchmark)